default export fix for NewsFeedScreen
